Move valid layouts to an external independent module
https://coveord.atlassian.net/browse/JSUI-1951
