+++ src/ui/ResultLayoutSelector/ValidLayout.ts
@@ -1,0 +1,6 @@
+/**
+ * The possible valid and supported layouts.
+ *
+ * See the [Result Layouts](https://developers.coveo.com/x/yQUvAg) documentation.
+ */
+export type ValidLayout = 'list' | 'card' | 'table';
