+++ src/ui/Templates/TemplateFromAScriptTag.ts
@@ -2,14 +2,14 @@ import { Template, IFieldsToMatch, TemplateRole } from './Template';
 import { Utils } from '../../utils/Utils';
 import { TemplateConditionEvaluator } from './TemplateConditionEvaluator';
 import { ComponentOptions, IComponentOptionsFieldsOption } from '../Base/ComponentOptions';
-import ResultLayoutSelectorModule = require('../ResultLayoutSelector/ResultLayoutSelector');
 import { $$, Dom } from '../../utils/Dom';
 import * as _ from 'underscore';
 import { Initialization } from '../Base/Initialization';
+import { ValidLayout } from '../ResultLayoutSelector/ValidLayout';
 
 export interface ITemplateFromStringProperties {
   condition?: string;
-  layout?: ResultLayoutSelectorModule.ValidLayout;
+  layout?: ValidLayout;
   mobile?: boolean;
   tablet?: boolean;
   desktop?: boolean;
@@ -107,9 +107,9 @@ export class TemplateFromAScriptTag {
     return Utils.parseBooleanIfNotUndefined(this.scriptTag.getAttribute(attribute));
   }
 
-  parseLayout(): ResultLayoutSelectorModule.ValidLayout {
+  parseLayout(): ValidLayout {
     const layout = this.scriptTag.getAttribute('data-layout');
-    return <ResultLayoutSelectorModule.ValidLayout>layout;
+    return <ValidLayout>layout;
   }
 
   static fromString(
